pyproject.toml
[project]
name = "lunch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "scikit-image>=0.25.2",
]

[dependency-groups]
dev = [
    "opencv-python>=4.11.0.86",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
]
